DEV-5348: Rework command to re-use existing method
